refactor: Movement module for new architecture
