+++ src/modules/Warehouse/views/Movement/CreateOrEdit/components/FormMovement/components/FormMovementDetail.vue
