+++ src/modules/Warehouse/views/Movement/CreateOrEdit/components/FormMovement/components/TableSearchDetail.vue
