+++ src/modules/Warehouse/views/Movement/CreateOrEdit/components/FormMovement/composables/useMovementDetail.ts
