+++ src/modules/Warehouse/views/Movement/Datagrid/composables/useIndex.ts
@@ -2,7 +2,7 @@ import { computed, reactive, onMounted } from "vue"
 import { onBeforeRouteUpdate, useRoute } from "vue-router"
 import useTableGrid from "@/core/composables/useTableGrid"
 import useHttp from "@/core/composables/useHttp"
-import MovementService from "../../services/Movement"
+import MovementService from "@/modules/Warehouse/services/Movement"
 
 type Params =  string | string[][] | Record<string, string> | URLSearchParams | undefined
